+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Fixed inaccurate CSI Snapshot logs. (Issue [#629](https://github.com/NetApp/trident/issues/629))
 - Fixed issue with unpublishing volumes on deleted nodes. (Issue [#691](https://github.com/NetApp/trident/issues/691))
 - Added handling of filesystem inconsistencies on block devices. (Issue [#656](https://github.com/NetApp/trident/issues/656))
+- Fixed issue where ANF driver failed to clone a volume with multiple export rules.
 
 **Enhancements**